+++ docs/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+
+
+## 1.12.0-alpha.1 - 2020-03-20
 ### Fixed
 * Avoid extra re-renders on WeekView
 * Minor style fixes in WeekView